01-docker-terraform/docker-sql/pyproject.toml
[project]
name = "dtc-etl"
version = "0.1.0"
description = "DTC DE Zoomcamp - high performance ingestion pipeline"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "loguru>=0.7.2",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "SQLAlchemy>=2.0.0",
  "psycopg2-binary>=2.9.9",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["dtc_etl*"]

# Optional: dev tools (kalau kamu mau)
[dependency-groups]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "pytest>=8.0.0",
]
